+++ src/wormbase/names/entity.clj
@@ -6,6 +6,7 @@
    [datomic.api :as d]
    [compojure.api.sweet :as sweet]
    [datomic.api :as d]
+   [magnetcoop.stork :as stork]
    [ring.util.http-response :refer [bad-request!
                                     conflict!
                                     conflict
@@ -345,16 +346,14 @@
   ([conn id-ident id-template prov generic? enabled? name-required?]
    (let [attrs-tx-data (conj (generic-attrs id-ident) prov)
          counter-ident (keyword "counter" (namespace id-ident))
-         tx-res @(d/transact-async conn attrs-tx-data)
-         temp-id (str id-ident)
          reg-tx-data [[:db/add id-ident :wormbase.names/id-template-format id-template]
                       [:db/add id-ident :wormbase.names/entity-type-generic? generic?]
                       [:db/add id-ident :wormbase.names/entity-type-enabled? enabled?]
                       [:db/add id-ident :wormbase.names/name-required? name-required?]
                       [:db/add counter-ident counter-ident 0N]]]
-     (when (:db-after tx-res)
-       @(d/transact-async conn reg-tx-data)
-       attrs-tx-data)))
+     (stork/ensure-installed conn {:id id-ident :tx-data attrs-tx-data})
+     (stork/ensure-installed conn {:id (keyword (namespace id-ident) "generic-registry")
+                                   :tx-data reg-tx-data})))
   ([conn id-ident id-template prov]
    (register-entity-schema conn id-ident id-template prov true true true)))